+++ test/ambiguities.jl
@@ -4,6 +4,6 @@
 
     # Only run on newer Julia versions.
     if VERSION > v"1.6"
-        @test isempty(Test.detect_ambiguities(Base, Core, CommonMark; recursive = true))
+        @test isempty(Test.detect_ambiguities(CommonMark))
     end
 end